app/requirements.txt
django==4.2
django-cors-headers==4.1.0
django-debug-toolbar==4.1.0
django_split_settings==1.2.0
django_postgres_extra==2.0.8
django-extensions==3.2.3
django-tinymce==3.6.1
django-stubs[compatible-mypy]==4.2.1
psycopg2==2.9.5

fastapi==0.92.0
orjson==3.9.1
dependency-injector==4.41.0

uvicorn[standard]==0.21.1
gunicorn==20.1.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"

python-jose[cryptography]==3.3.0